% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpng.R
\name{ggpng}
\alias{ggpng}
\alias{ggPNG}
\alias{theme_transparent}
\title{Create a picture with a transparent background.}
\usage{
ggpng(x, dp = 600, ...)

ggPNG(x, p, dp = 600, ...)

theme_transparent()
}
\arguments{
\item{x}{A file name that does not have a suffix and is entered as a string.}

\item{dp}{dpi = 600}

\item{...}{see:\link{ggsave}}

\item{p}{ggplot2 object}
}
\description{
Use the Imagemagick command line to convert the PDF saved by ggplot2 to PNG format with a transparent background and to set the resolution.
}
\author{
tsia
}
